+++ zarrs_object_store/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Move integration tests into `tests/`
+
 ## [0.4.0] - 2025-03-10
 
 ### Changed